add ground bounce example
